+++ packages/node/src/handlers.ts
@@ -1,7 +1,6 @@
 import { captureException, getCurrentHub } from '@sentry/core';
 import { Event } from '@sentry/types';
 import { forget, isString, logger, normalize } from '@sentry/utils';
-import { flush } from './sdk';
 import * as cookie from 'cookie';
 import * as domain from 'domain';
 import * as http from 'http';
@@ -9,6 +8,7 @@ import * as os from 'os';
 import * as url from 'url';
 
 import { NodeClient } from './client';
+import { flush } from './sdk';
 
 const DEFAULT_SHUTDOWN_TIMEOUT = 2000;
 
@@ -230,12 +230,17 @@ export function requestHandler(options?: {
     next: (error?: any) => void,
   ): void {
     if (options && options.flushTimeout && options.flushTimeout > 0) {
-      const _end = res.end
-
-      res.end = async function end (chunk?: any, encodingOrCb?: string | Function, cb?: Function) {
-        await flush(options.flushTimeout)
-        return _end.call(this, chunk, encodingOrCb, cb)
-      }
+      // tslint:disable-next-line: no-unbound-method
+      const _end = res.end;
+      res.end = function(chunk?: any | (() => void), encoding?: string | (() => void), cb?: () => void): void {
+        flush(options.flushTimeout)
+          .then(() => {
+            _end.call(this, chunk, encoding, cb);
+          })
+          .catch(e => {
+            logger.error(e);
+          });
+      };
     }
     const local = domain.create();
     local.add(req);